% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBrM.R
\name{plotBrM}
\alias{plotBrM}
\title{Plot Brownian motion}
\usage{
plotBrM(x, pointcolor = c("springgreen", "violetred"), ...)
}
\arguments{
\item{x}{matrix with two columns}

\item{pointcolor}{color of start and end points}

\item{...}{passed to plot()}
}
\value{
None
}
\description{
Plot Brownian motion
}
\examples{
x <- simBrM(1000)
plotBrM(x)
}
