% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBrM-package.R
\docType{package}
\name{simBrM-package}
\alias{simBrM}
\alias{simBrM-package}
\title{simBrM: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmj3197/simBrM}
  \item Report bugs at \url{https://github.com/rmj3197/simBrM/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
