% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBrM.R
\name{simBrM}
\alias{simBrM}
\title{Simulate Brownian motion}
\usage{
simBrM(n, sigma = 1)
}
\arguments{
\item{n}{number of steps}

\item{sigma}{SD of step size}
}
\value{
an nx2 matrix of positions
}
\description{
Simulate Brownian motion
}
\examples{
x <- simBrM(1000)
}
